refactor(tags): tag-service をドメイン単位の service に分割する
tag-service.ts（895行）を facade 化し、実装を責務ごとに移動:

- tag-mutation-service.ts: create / update
- tag-merge-service.ts: merge
- tag-delete-service.ts: delete / deleteGroup
- tag-group-service.ts: renameGroup / ungroupTags
- tag-sort-order.ts: getNextSortOrder / makeRoomAtTop（共有）
- tag-association-strategy.ts: countTagAssociations / applyTagStrategy（共有）
- tag-row-transform.ts: transformDbTag（共有）

ロジックは書き換えずに移動のみ。router からの呼び出しシグネチャは不変。
テストも同じ軸で 5 ファイルへ分割（facade 経由のまま、mock ヘルパーは
tag-service-test-helpers.ts へ集約）。全ファイル 400 行未満。

Refs #1529
